+++ stac_pydantic/extensions.py
@@ -1,4 +1,5 @@
 import json
+from functools import lru_cache
 from typing import Any, Dict, Union
 
 import jsonschema
@@ -9,10 +10,21 @@ from stac_pydantic.collection import Collection
 from stac_pydantic.item import Item
 
 
+@lru_cache(maxsize=128)
+def _fetch_and_cache_schema(url: str) -> dict:
+    """Fetch the remote JSON schema, if not already cached."""
+    req = requests.get(url)
+    return req.json()
+
+
 def validate_extensions(
     stac_obj: Union[Item, Collection, Catalog, Dict[str, Any]],
     reraise_exception: bool = False,
 ) -> bool:
+    """
+    Fetch the remote JSON schema, if not already cached, and validate the STAC
+    object against that schema.
+    """
     if isinstance(stac_obj, dict):
         stac_dict = stac_obj
     else:
@@ -23,8 +35,7 @@ def validate_extensions(
     try:
         if stac_dict["stac_extensions"]:
             for ext in stac_dict["stac_extensions"]:
-                req = requests.get(ext)
-                schema = req.json()
+                schema = _fetch_and_cache_schema(ext)
                 jsonschema.validate(instance=stac_dict, schema=schema)
     except Exception:
         if reraise_exception: